fix: Enable all features in dev mode and eliminate duplication
- Add context menu, command palette, global menu bar to dev route
- Extract action registration to shared registerDefaultActions() function
- Both table and dev routes now use identical action/UI code
- Fixes: flipping, context menu, and action handle now work in dev mode
